pyproject.toml
[tool.poetry]
name = "ncku"
version = "0.1.0"
description = ""
authors = ["Kiwi_Lin <opda0887@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.10.4"
sqlalchemy = "^2.0.12"
psycopg2-binary = "^2.9.6"
fastapi = "^0.95.1"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
requests = "^2.30.0"
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.0"
pyjwt = "^2.6.0"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
